requirements.txt
llama-index
openai
python-dotenv
atprototools
easydict
pytz
python-dateutil
cairosvg